backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytz==2023.3
python-multipart==0.0.9
argon2-cffi==23.1.0
email-validator==2.1.0
